requirements.txt
-i https://pypi.org/simple
asgiref==3.3.4; python_version >= '3.6'
django-safedelete==1.0.0
django==3.2.4
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1; python_version >= '3.5'
